pyproject.toml
[tool.poetry]
name = "training-the-lora-adapter-for-the-diffusion-model"
version = "0.1.0"
description = "LoRA training for Stable Diffusion"
authors = ["Panchenko Vasilii vaigpanchenko@edu.hse.ru"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.9"
clip = { git = "https://github.com/openai/CLIP.git" }

# Core ML
torch = "2.0.1"
torchvision = "0.15.2"
diffusers = "0.20.2"
transformers = "4.31.0"
accelerate = "0.21.0"
peft = "0.6.2"
huggingface-hub = "0.17.3"

# Data & Models
datasets = "2.14.4"
safetensors = "0.3.3"
ftfy = "6.1.1"
pyarrow = "12.0.1"

# Config & Utils
hydra-core = "1.3.0"
omegaconf = "2.3.0"
Pillow = "9.5.0"
tqdm = "4.65.0"
matplotlib = "3.7.2"
numpy = "1.24.3"

# Metrics
openai = "0.28.1"
openai-clip = "1.0.1"
sentence-transformers = "2.2.2"

# Logging
clearml = "1.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pre-commit (>=3.7.1,<4.0.0)",
    "ruff (>=0.14.10,<0.15.0)"
]
